requirements.txt
aiohttp
lxml
pyyaml
ratelimit